Use binstall
