+++ installers/npm/.npmignore
@@ -1,0 +1,3 @@
+README.md
+.gitignore
+.git
